requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
cryptography==3.2.1       # via josepy, mozilla-django-oidc, pyopenssl
django-crispy-forms==1.9.2  # via -r requirements.in
django==3.1.2             # via -r requirements.in, mozilla-django-oidc
idna==2.10                # via requests
josepy==1.4.0             # via mozilla-django-oidc
mozilla-django-oidc==1.2.4  # via -r requirements.in
oauthlib==3.1.0           # via requests-oauthlib
pycparser==2.20           # via cffi
pyopenssl==19.1.0         # via josepy
pytz==2020.1              # via django
pyyaml==5.3.1             # via -r requirements.in
requests-oauthlib==1.3.0  # via -r requirements.in
requests==2.24.0          # via mozilla-django-oidc, requests-oauthlib
six==1.15.0               # via cryptography, josepy, mozilla-django-oidc, pyopenssl
sqlparse==0.4.1           # via django
urllib3==1.25.11          # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
